refactor: lift TestOps provider base
