+++ src/dev_health_ops/connectors/testops/github_actions.py
@@ -6,7 +6,7 @@ from uuid import UUID
 
 from dev_health_ops.metrics.testops_schemas import JobRunRow, PipelineRunExtendedRow
 
-from .base import BasePipelineAdapter, PipelineSyncBatch
+from dev_health_ops.providers._base import BasePipelineAdapter, PipelineSyncBatch
 
 
 class GitHubActionsAdapter(BasePipelineAdapter):
